Refactor userCmdHandler
